- refactored getIcon function
