+++ modules/logging/ui/src/sap/commerce/toolset/logging/actionSystem/ApplyBundledTemplateAction.kt
@@ -1,0 +1,54 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package sap.commerce.toolset.logging.actionSystem
+
+import com.intellij.openapi.actionSystem.ActionPlaces
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.util.asSafely
+import sap.commerce.toolset.HybrisIcons
+import sap.commerce.toolset.logging.CxLoggerAccess
+import sap.commerce.toolset.logging.ui.tree.nodes.BundledLoggersTemplateItemNode
+
+class ApplyBundledTemplateAction : AnAction() {
+
+    override fun getActionUpdateThread() = ActionUpdateThread.BGT
+
+    override fun actionPerformed(e: AnActionEvent) {
+        e.presentation.isVisible = ActionPlaces.ACTION_SEARCH != e.place
+        if (!e.presentation.isVisible) return
+
+        val project = e.project ?: return
+
+        e.selectedNode()
+            ?.asSafely<BundledLoggersTemplateItemNode>()
+            ?.loggers
+            ?.let {
+                CxLoggerAccess.Companion.getInstance(project).setLoggers(it) { _, result ->
+                    println(result)
+                }
+            }
+    }
+
+    override fun update(e: AnActionEvent) {
+        e.presentation.text = "Apply Template"
+        e.presentation.icon = HybrisIcons.Log.Template.EXECUTE
+    }
+}
