+++ modules/logging/ui/src/sap/commerce/toolset/logging/actionSystem/CxLoggerActions.kt
@@ -21,7 +21,6 @@ package sap.commerce.toolset.logging.actionSystem
 import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.project.DumbAware
 import com.intellij.util.asSafely
-import sap.commerce.toolset.HybrisIcons
 import sap.commerce.toolset.Notifications
 import sap.commerce.toolset.logging.CxLoggerAccess
 import sap.commerce.toolset.logging.CxLoggersConstants
@@ -71,32 +70,6 @@ class WarnLoggerAction : CxLoggerAction(LogLevel.WARN)
 class ErrorLoggerAction : CxLoggerAction(LogLevel.ERROR)
 class FatalLoggerAction : CxLoggerAction(LogLevel.FATAL)
 
-class ApplyBundledTemplateAction : AnAction() {
-
-    override fun getActionUpdateThread() = ActionUpdateThread.BGT
-
-    override fun actionPerformed(e: AnActionEvent) {
-        e.presentation.isVisible = ActionPlaces.ACTION_SEARCH != e.place
-        if (!e.presentation.isVisible) return
-
-        val project = e.project ?: return
-
-        e.selectedNode()
-            ?.asSafely<BundledLoggersTemplateItemNode>()
-            ?.loggers
-            ?.let {
-                CxLoggerAccess.getInstance(project).setLoggers(it) { _, result ->
-                    println(result)
-                }
-            }
-    }
-
-    override fun update(e: AnActionEvent) {
-        e.presentation.text = "Apply Template"
-        e.presentation.icon = HybrisIcons.Log.Template.EXECUTE
-    }
-}
-
 class CxLoggersContextMenuActionGroup : ActionGroup(), DumbAware {
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
 
@@ -126,7 +99,7 @@ class CxLoggersContextMenuActionGroup : ActionGroup(), DumbAware {
     }
 }
 
-private fun AnActionEvent.selectedNode(): Any? = this.getData(PlatformCoreDataKeys.CONTEXT_COMPONENT)
+ fun AnActionEvent.selectedNode(): Any? = this.getData(PlatformCoreDataKeys.CONTEXT_COMPONENT)
     ?.asSafely<LoggersOptionsTree>()
     ?.selectionPath
     ?.lastPathComponent
